requirements.txt
fastapi
uvicorn[standard]
python-multipart
jinja2
PyPDF2
python-docx
chromadb
google-genai
python-dotenv
